% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_midi_file.R
\name{is_midi_file}
\alias{is_midi_file}
\title{checks if file is a midi file}
\usage{
is_midi_file(file)
}
\arguments{
\item{file}{path of the midi file to be played}
}
\value{
logical
}
\description{
Check if file is a midi file
}
