% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midiplayR-package.R
\docType{package}
\name{midiplayR-package}
\alias{midiplayR}
\alias{midiplayR-package}
\title{midiplayR: Play MIDI files from within R}
\description{
Play midi files on Mac, Linux, and Windows. Checks if MIDI player software is installed and recommends installing platform specific player software if not.
}
\author{
\strong{Maintainer}: Erich Neuwirth \email{erich.neuwirth@univie.ac.at}

}
\keyword{internal}
