% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_midi_file.R
\name{play_midi_file}
\alias{play_midi_file}
\title{play MIDI file}
\usage{
play_midi_file(file)
}
\arguments{
\item{file}{path of the midi file to be played}
}
\description{
`play_midi_file` takes a midi file and plays it
through the default system audio output.

On Mac and Lixux, a command line program playing MIDI files
need to be installed and the path set in the envorinment
variable `R_MIDIPLAYER`
}
\examples{
\dontrun{
play_midi_file(system.file("midifiles/Nokia_Tune.mid",package="midiplayR"))
}
}
